Add `GetCost(e)` to return the cost of `e` in O(1)
Add the method `IntMap.GetCost(elem)` which returns the last cost that was assigned to `elem`, no matter if `elem` is still contained in IntMap or not. This last part is important as it allows one to effectively use IntMap as an array with O(1) access to the value of index `elem`.